@@ -1,9 +1,11 @@
 package ru.kazakova_net.tourguideapp.activity;
 
 import android.os.Bundle;
-import android.support.design.widget.TabLayout;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.viewpager.widget.ViewPager;
+
+import com.google.android.material.tabs.TabLayout;
 
 import ru.kazakova_net.tourguideapp.R;
 import ru.kazakova_net.tourguideapp.adapter.AttractionPagerAdapter;
